chore: Renames and refactor to SearchInputField
