--- packages/components/src/molecules/SearchInput/index.ts
@@ -1,2 +1,0 @@
-export { default } from './SearchInput'
-export type { SearchInputProps, SearchInputRef } from './SearchInput'
