+++ packages/components/src/molecules/SearchInputField/index.ts
@@ -1,0 +1,5 @@
+export { default } from './SearchInputField'
+export type {
+  SearchInputFieldProps,
+  SearchInputFieldRef,
+} from './SearchInputField'
